Hotfix 2

+++ src/main/java/ru/yandex/practicum/filmorate/service/DirectorService.java
@@ -2,7 +2,8 @@ package ru.yandex.practicum.filmorate.service;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import ru.yandex.practicum.filmorate.exception.DirectorNotFoundException;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Director;
 import ru.yandex.practicum.filmorate.storage.DirectorStorage;
 
@@ -19,35 +20,46 @@ public class DirectorService {
 
     public Director createDirector(Director director) {
         int generatedId = directorStorage.save(director);
-        Director createdDirector = directorStorage.getById(generatedId).orElse(null);
+        Director createdDirector = directorStorage.findById(generatedId).orElse(null);
 
         log.info("Добавлен режиссёр: {}", createdDirector);
         return createdDirector;
     }
 
     public Director updateDirector(Director director) {
+        Integer directorId = director.getId();
+        if (directorId == null) {
+            throw new ValidationException("Требуется корректный id параметр");
+        } else if (!directorStorage.existsById(directorId)) {
+            throw new NotFoundException("Режиссёр с id " + directorId + " не найден");
+        }
+
         directorStorage.update(director);
-        Director updatedDirector = directorStorage.getById(director.getId()).orElse(null);
+        Director updatedDirector = directorStorage.findById(directorId).orElse(null);
 
         log.info("Обновлён режиссёр: {}", updatedDirector);
         return updatedDirector;
     }
 
-    public List<Director> findAllDirectors() {
+    public List<Director> getAllDirectors() {
         log.info("Возвращен список всех режиссёров");
-        return directorStorage.getAll();
+        return directorStorage.findAll();
     }
 
-    public Director findDirectorById(Integer directorId) {
-        Director director = directorStorage.getById(directorId)
-                .orElseThrow(() -> new DirectorNotFoundException("Режиссёр с id " + directorId + " не найден"));
+    public Director getDirectorById(Integer directorId) {
+        Director director = directorStorage.findById(directorId)
+                .orElseThrow(() -> new NotFoundException("Режиссёр с id " + directorId + " не найден"));
 
         log.info("Получен режиссёр: {}", director);
         return director;
     }
 
-    public boolean deleteDirector(Integer directorId) {
+    public void deleteDirector(Integer directorId) {
+        if (!directorStorage.existsById(directorId)) {
+            throw new NotFoundException("Режиссёр с id " + directorId + " не найден");
+        }
+
         log.info("Режиссёр с id {} удалён.", directorId);
-        return directorStorage.delete(directorId);
+        directorStorage.delete(directorId);
     }
 }
