+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -4,13 +4,15 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
-import ru.yandex.practicum.filmorate.exception.FilmNotFoundException;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Director;
 import ru.yandex.practicum.filmorate.model.Event;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.FilmDirector;
 import ru.yandex.practicum.filmorate.storage.*;
 
+import javax.validation.constraints.NotNull;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -22,17 +24,20 @@ import static ru.yandex.practicum.filmorate.model.EventType.LIKE;
 @Service
 public class FilmService {
     private final FilmStorage filmStorage;
+    private final UserStorage userStorage;
     private final LikeStorage likeStorage;
     private final FilmDirectorStorage filmDirectorStorage;
     private final EventStorage eventStorage;
     private final DirectorStorage directorStorage;
 
     public FilmService(@Qualifier("filmDbStorage") FilmStorage filmStorage,
+                       UserStorage userStorage,
                        LikeStorage likeStorage,
                        FilmDirectorStorage filmDirectorStorage,
                        EventStorage eventStorage,
                        DirectorStorage directorStorage) {
         this.filmStorage = filmStorage;
+        this.userStorage = userStorage;
         this.likeStorage = likeStorage;
         this.filmDirectorStorage = filmDirectorStorage;
         this.eventStorage = eventStorage;
@@ -41,40 +46,57 @@ public class FilmService {
 
     public Film createFilm(Film film) {
         int generatedId = filmStorage.save(film);
-        Film createdFilm = filmStorage.getById(generatedId).orElse(null);
+        Film createdFilm = filmStorage.findById(generatedId).orElse(null);
 
         log.info("Добавлен фильм: {}", createdFilm);
         return createdFilm;
     }
 
     public Film updateFilm(Film film) {
+        Integer filmId = film.getId();
+        if (filmId == null) {
+            throw new ValidationException("Требуется корректный id параметр");
+        } else if (!filmStorage.existsById(filmId)) {
+            throw new NotFoundException("Фильм с id " + filmId + " не найден");
+        }
+
         filmStorage.update(film);
-        Film updatedFilm = filmStorage.getById(film.getId()).orElse(null);
+        Film updatedFilm = filmStorage.findById(film.getId()).orElse(null);
 
         log.info("Обновлён фильм: {}", updatedFilm);
         return updatedFilm;
     }
 
-    public List<Film> findAllFilms() {
+    public List<Film> getAllFilms() {
         log.info("Возвращен список всех фильмов");
-        return filmStorage.getAll();
+        return filmStorage.findAll();
     }
 
-    public Film findFilmById(Integer filmId) {
-        Film film = filmStorage.getById(filmId)
-                .orElseThrow(() -> new FilmNotFoundException("Фильм с id " + filmId + " не найден"));
+    public Film getFilmById(Integer filmId) {
+        Film film = filmStorage.findById(filmId)
+                .orElseThrow(() -> new NotFoundException("Фильм с id " + filmId + " не найден"));
 
         log.info("Получен фильм: {}", film);
         return film;
     }
 
-    public boolean deleteFilmById(Integer filmId) {
+    public void deleteFilmById(Integer filmId) {
+        if (!filmStorage.existsById(filmId)) {
+            throw new NotFoundException("Фильм с id " + filmId + " не найден");
+        }
+
         log.info("Фильм с id {} удален: ", filmId);
-        return filmStorage.deleteById(filmId);
+        filmStorage.deleteById(filmId);
     }
 
     public boolean likeFilm(Integer filmId, Integer userId) {
         try {
+            if (!filmStorage.existsById(filmId)) {
+                throw new NotFoundException("Фильм с id " + filmId + " не найден");
+            } else if (!userStorage.existsById(userId)) {
+                throw new NotFoundException("Пользователь с id " + userId + " не найден");
+            }
+
             likeStorage.save(filmId, userId);
             log.info("Пользователь с id {} поставил лайк фильму с id {}", userId, filmId);
 
@@ -93,27 +115,29 @@ public class FilmService {
         }
     }
 
-    public boolean dislikeFilm(Integer filmId, Integer userId) {
-        if (likeStorage.delete(filmId, userId)) {
-            log.info("Пользователь с id {} убрал лайк с фильма с id {}", userId, filmId);
-
-            Event event = Event.builder()
-                    .userId(userId)
-                    .entityId(filmId)
-                    .eventType(LIKE)
-                    .operation(REMOVE)
-                    .build();
-            eventStorage.save(event);
-
-            return true;
+    public void dislikeFilm(@NotNull Integer filmId, @NotNull Integer userId) {
+        if (!filmStorage.existsById(filmId)) {
+            throw new NotFoundException("Фильм с id " + filmId + " не найден");
+        } else if (!userStorage.existsById(userId)) {
+            throw new NotFoundException("Пользователь с id " + userId + " не найден");
         }
-        return false;
+
+        likeStorage.delete(filmId, userId);
+        log.info("Пользователь с id {} убрал лайк с фильма с id {}", userId, filmId);
+
+        Event event = Event.builder()
+                .userId(userId)
+                .entityId(filmId)
+                .eventType(LIKE)
+                .operation(REMOVE)
+                .build();
+        eventStorage.save(event);
     }
 
-    public List<Film> findTopFilms(Integer count) {
-        List<Film> topFilms = likeStorage.getPopularFilmsIds(count)
+    public List<Film> getTopFilms(Integer count) {
+        List<Film> topFilms = likeStorage.findPopularFilmsIds(count)
                 .stream()
-                .map(filmStorage::getById)
+                .map(filmStorage::findById)
                 .filter(Optional::isPresent)
                 .map(Optional::get)
                 .collect(Collectors.toList());
@@ -123,9 +147,13 @@ public class FilmService {
     }
 
     public List<Film> getRecommendations(Integer userId) {
-        List<Film> recommendedFilms = likeStorage.getRecommendFilmsIds(userId)
+        if (!userStorage.existsById(userId)) {
+            throw new NotFoundException("Пользователь с id " + userId + " не найден");
+        }
+
+        List<Film> recommendedFilms = likeStorage.findRecommendFilmsIds(userId)
                 .stream()
-                .map(filmStorage::getById)
+                .map(filmStorage::findById)
                 .filter(Optional::isPresent)
                 .map(Optional::get)
                 .collect(Collectors.toList());
@@ -134,10 +162,16 @@ public class FilmService {
         return recommendedFilms;
     }
 
-    public List<Film> findCommonFilms(Integer userId, Integer friendId) {
-        List<Film> films = likeStorage.getCommonFilmsIds(userId, friendId)
+    public List<Film> getCommonFilms(Integer userId, Integer friendId) {
+        if (!userStorage.existsById(userId)) {
+            throw new NotFoundException("Пользователь с id " + userId + " не найден");
+        } else if (!userStorage.existsById(friendId)) {
+            throw new NotFoundException("Пользователь с id " + friendId + " не найден");
+        }
+
+        List<Film> films = likeStorage.findCommonFilmsIds(userId, friendId)
                 .stream()
-                .map(filmStorage::getById)
+                .map(filmStorage::findById)
                 .filter(Optional::isPresent)
                 .map(Optional::get)
                 .collect(Collectors.toList());
@@ -146,10 +180,10 @@ public class FilmService {
         return films;
     }
 
-    public List<Film> findTopFilmsByGenreAndYear(Integer count, Integer genreId, Integer year) {
-        List<Film> topFilms = likeStorage.getPopularFilmsIds(count)
+    public List<Film> getTopFilmsByGenreAndYear(Integer count, Integer genreId, Integer year) {
+        List<Film> topFilms = likeStorage.findPopularFilmsIds(count)
                 .stream()
-                .map(filmStorage::getById)
+                .map(filmStorage::findById)
                 .filter(Optional::isPresent)
                 .map(Optional::get)
                 .collect(Collectors.toList());
@@ -162,11 +196,15 @@ public class FilmService {
         return topFilmsFiltered;
     }
 
-    public List<Film> findFilmsByDirector(Integer directorId, String sortBy) {
-        List<Film> directorFilms = filmDirectorStorage.getAllByDirector(directorId)
+    public List<Film> getFilmsByDirector(Integer directorId, String sortBy) {
+        if (!directorStorage.existsById(directorId)) {
+            throw new NotFoundException("Режиссёр с id " + directorId + " не найден");
+        }
+
+        List<Film> directorFilms = filmDirectorStorage.findAllByDirector(directorId)
                 .stream()
                 .map(FilmDirector::getFilmId)
-                .map(filmStorage::getById)
+                .map(filmStorage::findById)
                 .filter(Optional::isPresent)
                 .map(Optional::get)
                 .collect(Collectors.toList());
@@ -175,7 +213,7 @@ public class FilmService {
             directorFilms.sort(Comparator.comparing(Film::getReleaseDate));
         } else if (sortBy.equals("likes")) {
             directorFilms.sort((film1, film2) ->
-                    likeStorage.getAllByFilmId(film1.getId()).size() - likeStorage.getAllByFilmId(film2.getId()).size());
+                    likeStorage.findAllByFilmId(film1.getId()).size() - likeStorage.findAllByFilmId(film2.getId()).size());
         } else {
             throw new RuntimeException("Укажите параметр сортировки year/likes");
         }
@@ -184,49 +222,49 @@ public class FilmService {
         return directorFilms;
     }
 
-    public List<Film> findFilmsByFilmName(String query) {
-        List<Film> films = filmStorage.getAllByNameSubstring(query);
+    public List<Film> getFilmsByFilmName(String query) {
+        List<Film> films = filmStorage.findAllByNameSubstring(query);
 
         films.sort((film1, film2) ->
-                likeStorage.getAllByFilmId(film2.getId()).size() - likeStorage.getAllByFilmId(film1.getId()).size());
+                likeStorage.findAllByFilmId(film2.getId()).size() - likeStorage.findAllByFilmId(film1.getId()).size());
 
         log.info("Возвращен список найденных фильмов: {} ", films);
         return films;
     }
 
-    public List<Film> findFilmsByDirectorName(String query) {
+    public List<Film> getFilmsByDirectorName(String query) {
         List<Film> films = new ArrayList<>();
-        List<Director> directors = directorStorage.getAllByNameSubstring(query);
+        List<Director> directors = directorStorage.findAllByNameSubstring(query);
 
         for (Director director : directors) {
-            films.addAll(filmDirectorStorage.getAllByDirector(director.getId())
+            films.addAll(filmDirectorStorage.findAllByDirector(director.getId())
                     .stream()
                     .map(FilmDirector::getFilmId)
-                    .map(filmStorage::getById)
+                    .map(filmStorage::findById)
                     .filter(Optional::isPresent)
                     .map(Optional::get)
                     .collect(Collectors.toList()));
         }
 
         films.sort((film1, film2) ->
-                likeStorage.getAllByFilmId(film2.getId()).size() - likeStorage.getAllByFilmId(film1.getId()).size());
+                likeStorage.findAllByFilmId(film2.getId()).size() - likeStorage.findAllByFilmId(film1.getId()).size());
 
         log.info("Возвращен список найденных фильмов: {} ", films);
         return films;
     }
 
-    public List<Film> findFilmsByFilmNameAndDirectorName(String query) {
+    public List<Film> getFilmsByFilmNameAndDirectorName(String query) {
         Set<Film> films = new TreeSet<>((film1, film2) ->
-                likeStorage.getAllByFilmId(film2.getId()).size() - likeStorage.getAllByFilmId(film1.getId()).size());
-        films.addAll(filmStorage.getAllByNameSubstring(query));
+                likeStorage.findAllByFilmId(film2.getId()).size() - likeStorage.findAllByFilmId(film1.getId()).size());
+        films.addAll(filmStorage.findAllByNameSubstring(query));
 
-        List<Director> directors = directorStorage.getAllByNameSubstring(query);
+        List<Director> directors = directorStorage.findAllByNameSubstring(query);
 
         for (Director director : directors) {
-            films.addAll(filmDirectorStorage.getAllByDirector(director.getId())
+            films.addAll(filmDirectorStorage.findAllByDirector(director.getId())
                     .stream()
                     .map(FilmDirector::getFilmId)
-                    .map(filmStorage::getById)
+                    .map(filmStorage::findById)
                     .filter(Optional::isPresent)
                     .map(Optional::get)
                     .collect(Collectors.toList()));
